requirements.txt
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
Flask>=2.0.2